+++ Personas/BasePersona.py
@@ -2,6 +2,7 @@ import logging
 from typing import List, Tuple
 
 from AiOrchestration.AiOrchestrator import AiOrchestrator
+from Functionality.Knowledge import Knowledge
 from Functionality.Organising import Organising
 from Personas.PersonaSpecification.PersonaConstants import SELECT_WORKFLOW_INSTRUCTIONS
 from Utilities.ErrorHandler import ErrorHandler
@@ -84,12 +85,15 @@ class BasePersona:
         selected_files = Organising.get_relevant_files(user_messages)
         executor = AiOrchestrator(selected_files)
 
+        # ToDo: this will be called multiple times redundantly in a workflow, user_messages are small however and the
+        #  context can change from step to step so its not a priority
+        additional_context = Knowledge.search_encyclopedia(user_messages)
         # ToDo: How the application accesses and gives history to the llm will need to be optimised
         recent_history = [f"{entry[0]}: {entry[1]}" for entry in self.history[-self.MAX_HISTORY:]]
 
         try:
             output = executor.execute(
-                [self.instructions, self.configuration],
+                [self.instructions, self.configuration, additional_context],
                 user_messages,
                 assistant_messages=recent_history
             )